# Conflicts:
#	app/src/main/java/com/frontend/nutricheck/client/model/data_sources/data/MealItem.kt
#	app/src/main/java/com/frontend/nutricheck/client/ui/view/app_views/CreateMealPage.kt
#	app/src/main/java/com/frontend/nutricheck/client/ui/view/widgets/DishItemButton.kt
#	app/src/main/java/com/frontend/nutricheck/client/ui/view/widgets/ServingsPicker.kt

+++ app/src/main/java/com/frontend/nutricheck/client/model/data_sources/data/FoodComponent.kt
@@ -37,5 +37,5 @@ sealed interface FoodComponent {
     /**
      * Returns the servings of the food component
      */
-    val servings: Int
+    val servings: Double
 }
