+++ app/src/main/java/com/frontend/nutricheck/client/model/data_sources/data/FoodProduct.kt
@@ -9,6 +9,6 @@ data class FoodProduct(
     override val carbohydrates: Double = 0.0,
     override val protein: Double = 0.0,
     override val fat: Double = 0.0,
-    override val servings: Int = 1,
+    override val servings: Double = 1.0,
     val servingSize: ServingSize = ServingSize.ONEHOUNDREDGRAMS
 ) : FoodComponent
