+++ app/src/main/java/com/frontend/nutricheck/client/model/data_sources/persistence/entity/IngredientEntity.kt
@@ -34,6 +34,6 @@ data class IngredientEntity (
     val recipeId: String,
     val foodProductId: String,
     val quantity: Double,
-    val servings: Int,
+    val servings: Double,
     val servingSize: ServingSize
 )
