+++ app/src/main/java/com/frontend/nutricheck/client/model/data_sources/persistence/entity/MealFoodItemEntity.kt
@@ -34,6 +34,6 @@ data class MealFoodItemEntity(
     val mealId: String,
     val foodProductId: String,
     val quantity: Double,
-    val servings: Int,
+    val servings: Double,
     val servingSize: ServingSize
 )
